Added access control

+++ app/Http/Controllers/PostsController.php
@@ -8,6 +8,16 @@ use App\Post;
 class PostsController extends Controller
 {
     /**
+     * Create a new controller instance.
+     *
+     * @return void
+     */
+    public function __construct()
+    {
+        //prevent guests from doing CRUD operations
+        $this->middleware('auth', ['except' => ['index', 'show']]);
+    }
+    /**
      * Display a listing of the resource.
      *
      * @return \Illuminate\Http\Response
@@ -69,7 +79,11 @@ class PostsController extends Controller
      */
     public function edit($id)
     {
-        return view('posts.edit')->with('post', Post::find($id));
+        $post = Post::find($id);
+        if (Auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page!');
+        }
+        return view('posts.edit')->with('post', $post);
     }
 
     /**
@@ -103,6 +117,9 @@ class PostsController extends Controller
     public function destroy($id)
     {
         $post = Post::find($id);
+        if (Auth()->user()->id !== $post->user_id) {
+            return redirect('/posts')->with('error', 'Unauthorized Page!');
+        }
         $post->delete();
         return redirect('/posts')->with('success', 'Post Deleted');
     }
